web/requirements.txt
Django==1.9.6
gunicorn==19.3.0
django-crispy-forms==1.6.0
djangorestframework==3.3.3
django-pandas==0.4.1
pandas-datareader==0.2.1
blaze==0.10.1
drf-nested-routers==0.11.1
